+++ x8086NetEmu/Adapters/Serial/MouseAdapter.vb
@@ -21,7 +21,7 @@
         MyBase.New(cpu)
         If MyBase.CPU.PIC IsNot Nothing Then irq = MyBase.CPU.PIC.GetIrqLine(4)
 
-        For i As UInt32 = &H3F8 To &H3F8 + 7
+        For i As UInt16 = &H3F8 To &H3F8 + 7
             RegisteredPorts.Add(i)
         Next
     End Sub
